backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0

# 인증 시스템 의존성
aiomysql==0.2.0
bcrypt==4.1.3
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
email-validator==2.1.1

